# force specialization in sv_ecef_to_ecef